LayoutTests/webgpu/whlsl-oob-access-expected.txt

PASS resultsFloat32Array[0] is 2
PASS resultsFloat32Array[1] is 4
PASS resultsFloat32Array[2] is 6
PASS resultsFloat32Array[3] is 8
PASS resultsFloat32Array[4] is 5
PASS resultsFloat32Array[5] is 6
PASS resultsFloat32Array[6] is 7
PASS resultsFloat32Array[7] is 8
PASS successfullyParsed is true

TEST COMPLETE
